+++ tests/nexus_test.py
@@ -206,6 +206,25 @@ def test_nxobject_getitem_by_class(nxroot):
     assert set(nxroot['entry'][NXevent_data]) == {'events_0', 'events_1'}
 
 
+def test_nxobject_getitem_by_class_get_fields(nxroot):
+    nxroot['entry'].create_class('log', NXlog)
+    nxroot['entry'].create_class('events_0', NXevent_data)
+    nxroot['entry']['field1'] = sc.arange('event', 4.0, unit='ns')
+    nxroot['entry']['field2'] = sc.arange('event', 2.0, unit='ns')
+    assert list(nxroot[Field]) == []
+    assert set(nxroot['entry'][Field]) == {'field1', 'field2'}
+
+
+def test_nxobject_getitem_by_class_tuple(nxroot):
+    nxroot['entry'].create_class('log', NXlog)
+    nxroot['entry'].create_class('events_0', NXevent_data)
+    nxroot['entry'].create_class('events_1', NXevent_data)
+    nxroot['entry']['field1'] = sc.arange('event', 4.0, unit='ns')
+    assert set(nxroot['entry'][(NXlog,
+                                NXevent_data)]) == {'log', 'events_0', 'events_1'}
+    assert set(nxroot['entry'][(NXlog, Field)]) == {'log', 'field1'}
+
+
 def test_nxobject_dataset_items_are_returned_as_Field(nxroot):
     events = nxroot['entry'].create_class('events_0', NXevent_data)
     events['event_time_offset'] = sc.arange('event', 5)